The problem of distinguishing prime numbers from composite numbers and of resolving the latter into their prime factors is known to be one of the most important and useful in arithmetic

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,9 +6,9 @@
  */
 int main(void)
 {
-long long int n = 612852475143;
-long long int i = 2;
-long long int maxPrime = 1;
+long int n = 612852475143;
+long int i = 2;
+long int maxPrime = 1;
 
 while (n > 1)
 {
